refactor: move file utils to proper module
