+++ src/files/mod.rs
@@ -2,3 +2,4 @@ pub mod archives;
 pub mod datadirs;
 pub mod filesys;
 pub mod magic;
+pub mod utils;
