+++ src/files/utils.rs
@@ -1,0 +1,30 @@
+use std::path::Path;
+
+pub fn get_file_extension(archive_path: &Path) -> &str {
+    let filename = archive_path
+        .file_name()
+        .and_then(|f| f.to_str())
+        .unwrap_or_default();
+
+    // Handle multi-part extensions like .tar.gz, .tar.xz, .tar.bz2
+    if filename.ends_with(".tar.gz") {
+        return "tar.gz";
+    } else if filename.ends_with(".tar.xz") {
+        return "tar.xz";
+    } else if filename.ends_with(".tar.bz2") {
+        return "tar.bz2";
+    }
+
+    // For single extensions, use the standard method
+    archive_path
+        .extension()
+        .and_then(|s| s.to_str())
+        .unwrap_or_default()
+}
+
+pub fn get_file_name(archive_path: &Path) -> &str {
+    archive_path
+        .file_name()
+        .and_then(|f| f.to_str())
+        .unwrap_or_default()
+}
